Finished AI. Next, optimizations, docs, and code cleanup

+++ src/ailogic/movenode.cc
@@ -9,7 +9,7 @@ MoveNode::MoveNode(const int &move_index) : move_index_(move_index),
                                             heuristic_value_(0) {
 }
 
-void MoveNode::SetHeuristicValue(const int &heuristic_value) {
+void MoveNode::SetHeuristicValue(const double &heuristic_value) {
   heuristic_value_ = heuristic_value;
 }
 
@@ -17,7 +17,7 @@ int MoveNode::MoveIndex() const {
   return move_index_;
 }
 
-int MoveNode::HeuristicValue() const {
+double MoveNode::HeuristicValue() const {
   return heuristic_value_;
 }
